Move AuthenticationError (and its test) to core.py
* Moved AuthenticationError to core.py
* Standardized imports in tests

closes #105
